move app, links, header, footer, horizontal line, and container components to ts
