+++ src/App.tsx
@@ -6,13 +6,13 @@ import Footer from "./components/Footer";
 import styles from "./App.module.css";
 
 export default function App() {
-    return (
-        <div className={styles.Layout}>
-            <Header />
-            <div id="content">
-                <Outlet />
-            </div>
-            <Footer />
-        </div>
-    );
+  return (
+    <div className={styles.Layout}>
+      <Header />
+      <div id="content">
+        <Outlet />
+      </div>
+      <Footer />
+    </div>
+  );
 }
